// src/lib/schema.d.ts
export interface Database {
  public: {
    Tables: {
// ...
      distributor_user: {
        Row: {
          bank_account: string | null
          created_at: string | null
          id: string
          location_id: string | null
          nif: string | null
          user: string | null
        }
        Insert: {
          bank_account?: string | null
          created_at?: string | null
          id?: string
          location_id?: string | null
          nif?: string | null
          user?: string | null
        }
        Update: {
          bank_account?: string | null
          created_at?: string | null
          id?: string
          location_id?: string | null
          nif?: string | null
          user?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "distributor_user_location_id_fkey"
            columns: ["location_id"]
            referencedRelation: "profile_location"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "distributor_user_user_fkey"
            columns: ["user"]
            referencedRelation: "users"
            referencedColumns: ["id"]
          }
        ]
      }
// ...
      users: {
        Row: {
          avatar_url: string | null
          bg_url: string | null
          birthdate: string | null
          cp_organizer_status: number | null
          created_at: string | null
          email: string | null
          id: string
          lastname: string | null
          name: string | null
          role: string | null
          updated_at: string | null
          username: string | null
        }
        Insert: {
          avatar_url?: string | null
          bg_url?: string | null
          birthdate?: string | null
          cp_organizer_status?: number | null
          created_at?: string | null
          email?: string | null
          id: string
          lastname?: string | null
          name?: string | null
          role?: string | null
          updated_at?: string | null
          username?: string | null
        }
        Update: {
          avatar_url?: string | null
          bg_url?: string | null
          birthdate?: string | null
          cp_organizer_status?: number | null
          created_at?: string | null
          email?: string | null
          id?: string
          lastname?: string | null
          name?: string | null
          role?: string | null
          updated_at?: string | null
          username?: string | null
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      set_claim: {
        Args: {
          uid: string
          claim: string
          value: Json
        }
        Returns: string
      }
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}